Корректный рендер офферов с сервера
