--- project/src/mocks/offers.js
@@ -1,630 +1,0 @@
-import {OfferType, City, CityName, ZOOM_OFFER, Good, HostName} from './constants.js';
-
-const offers = [
-  {
-    id: 'offer001',
-    title: 'Beautiful & luxurious apartment at great location',
-    description: 'A new spacious villa, one floor. All commodities, jacuzzi and beautiful scenery. Ideal for families or friends.',
-    type: OfferType.APARTMENT,
-    price: 120,
-    previewImage: 'img/apartment-01.jpg',
-    images: [
-      'img/apartment-01.jpg',
-      'img/apartment-02.jpg',
-    ],
-    rating: 5,
-    bedrooms: 3,
-    maxAdults: 6,
-    goods: [
-      Good.BABY_SEAT,
-      Good.CABEL_TV,
-      Good.COFFEE_MASHINE,
-      Good.DISHWASHER,
-    ],
-    host: {
-      id: 25,
-      name: HostName.ANGELINA,
-      isPro: true,
-      avatarUrl: 'img/avatar-angelina.jpg',
-    },
-    city: City[CityName.AMSTERDAM],
-    isPremium: true,
-    isFavourite: false,
-    location: {
-      latitude: 52.3909553943508,
-      longitude: 4.85309666406198,
-      zoom: ZOOM_OFFER,
-    },
-  },
-  {
-    id: 'offer002',
-    title: 'Wood and stone place',
-    description: 'Relax, rejuvenate and unplug in this ultimate rustic getaway experience in the country. In our beautiful screened Pondhouse, you can gaze at the stars and listen to the sounds of nature from your cozy warm bed.',
-    type: OfferType.PRIVATE_ROOM,
-    price: 80,
-    previewImage: 'img/room.jpg',
-    images: [
-      'img/studio-01.jpg',
-      'img/room.jpg',
-      'img/apartment-03.jpg',
-    ],
-    rating: 2.5,
-    bedrooms: 2,
-    maxAdults: 7,
-    goods: [
-      Good.WI_FI,
-      Good.KITCHEN,
-      Good.TOWELS,
-      Good.FRIDGE,
-      Good.CABEL_TV,
-    ],
-    host: {
-      id: 24,
-      name: HostName.MAX,
-      isPro: false,
-      avatarUrl: 'img/avatar-max.jpg',
-    },
-    city: City[CityName.AMSTERDAM],
-    isPremium: false,
-    isFavourite: true,
-    location: {
-      latitude: 52.369553943508,
-      longitude: 4.85309666406198,
-      zoom: ZOOM_OFFER,
-    },
-  },
-  {
-    id: 'offer003',
-    title: 'Canal View Prinsengracht',
-    description: 'I am happy to welcome you to my apartment in the city center! Three words: location, cosy and chic!',
-    type: OfferType.APARTMENT,
-    price: 132,
-    previewImage: 'img/apartment-02.jpg',
-    images: [
-      'img/apartment-03.jpg',
-      'img/apartment-01.jpg',
-      'img/apartment-02.jpg',
-    ],
-    rating: 4,
-    bedrooms: 3,
-    maxAdults: 5,
-    goods: [
-      Good.WI_FI,
-      Good.CABEL_TV,
-    ],
-    host: {
-      id: 25,
-      name: HostName.ANGELINA,
-      isPro: true,
-      avatarUrl: 'img/avatar-angelina.jpg',
-    },
-    city: City[CityName.AMSTERDAM],
-    isPremium: false,
-    isFavourite: false,
-    location: {
-      latitude: 52.3909553943508,
-      longitude: 4.929309666406198,
-      zoom: ZOOM_OFFER,
-    },
-  },
-  {
-    id: 'offer004',
-    title: 'Nice, cozy, warm big bed apartment',
-    description: 'Peaceful studio in the most wanted area in town. Quiet house Near of everything. Completely renovated. Lovely neighbourhood, lot of trendy shops, restaurants and bars in a walking distance.',
-    type: OfferType.APARTMENT,
-    price: 180,
-    previewImage: 'img/apartment-03.jpg',
-    images: [
-      'img/apartment-01.jpg',
-      'img/apartment-02.jpg',
-      'img/apartment-03.jpg',
-      'img/studio-01.jpg',
-    ],
-    rating: 3.5,
-    bedrooms: 4,
-    maxAdults: 5,
-    goods: [
-      Good.WI_FI,
-      Good.KITCHEN,
-      Good.TOWELS,
-      Good.FRIDGE,
-      Good.CABEL_TV,
-      Good.BABY_SEAT,
-      Good.HEATING,
-    ],
-    host: {
-      id: 25,
-      name: HostName.ANGELINA,
-      isPro: true,
-      avatarUrl: 'img/avatar-angelina.jpg',
-    },
-    city: City[CityName.AMSTERDAM],
-    isPremium: true,
-    isFavourite: true,
-    location: {
-      latitude: 52.3809553943508,
-      longitude: 4.939309666406198,
-      zoom: ZOOM_OFFER,
-    },
-  },
-  {
-    id: 'offer005',
-    title: 'Wood and stone place',
-    description: 'Discover daily local life in city center, friendly neighborhood, clandestine casino, karaoke, old-style artisans, art gallery and artist studio downstairs.',
-    type: OfferType.PRIVATE_ROOM,
-    price: 80,
-    previewImage: 'img/room.jpg',
-    images: [
-      'img/apartment-03.jpg',
-      'img/studio-01.jpg',
-      'img/room.jpg',
-      'img/apartment-02.jpg',
-      'img/apartment-01.jpg',
-    ],
-    rating: 1.9,
-    bedrooms: 1,
-    maxAdults: 2,
-    goods: [
-      Good.COFFEE_MASHINE,
-    ],
-    host: {
-      id: 23,
-      name: HostName.JAMES,
-      isPro: false,
-      avatarUrl: '',
-    },
-    city: City[CityName.AMSTERDAM],
-    isPremium: false,
-    isFavourite: false,
-    location: {
-      latitude: 52.369553943508,
-      longitude: 4.939309666406198,
-      zoom: ZOOM_OFFER,
-    },
-  },
-  {
-    id: 'offer006',
-    title: 'Torzimmer auf Schloss Arenfels',
-    description: 'Sie mieten unserer Torzimmer im Ostflüel des Schlosses. Das Schloss blickt auf eine 800 Jahre alte Geschichte zurück und ist DAS Wahrzeichens Bad Hönningens. Schloss Arenfels liegt mitten in den Weinbergen. Der Wald, die Weinberge und der Rhein laden zum spazieren und wandern ein.',
-    type: OfferType.APARTMENT,
-    price: 190,
-    previewImage: 'img/apartment-03.jpg',
-    images: [
-      'img/apartment-03.jpg',
-      'img/studio-01.jpg',
-      'img/room.jpg',
-      'img/apartment-02.jpg',
-      'img/apartment-01.jpg',
-    ],
-    rating: 4.8,
-    bedrooms: 4,
-    maxAdults: 5,
-    goods: [
-      Good.COFFEE_MASHINE,
-      Good.WI_FI,
-      Good.KITCHEN,
-      Good.TOWELS,
-      Good.FRIDGE,
-      Good.CABEL_TV,
-      Good.BABY_SEAT,
-      Good.HEATING,
-    ],
-    host: {
-      id: 23,
-      name: HostName.JAMES,
-      isPro: false,
-      avatarUrl: '',
-    },
-    city: City[CityName.HAMBURG],
-    isPremium: false,
-    isFavourite: true,
-    location: {
-      latitude: 53.554,
-      longitude: 9.998,
-      zoom: ZOOM_OFFER,
-    },
-  },
-  {
-    id: 'offer007',
-    title: 'Herrliche Aussicht auf den Fluss und die Umgebung',
-    description: 'Das Schloss blickt auf eine 800 Jahre alte Geschichte zurück und ist DAS Wahrzeichens Bad Hönningens. Schloss Arenfels liegt mitten in den Weinbergen. Der Wald, die Weinberge und der Rhein laden zum spazieren und wandern ein.',
-    type: OfferType.PRIVATE_ROOM,
-    price: 125,
-    previewImage: 'img/apartment-02.jpg',
-    images: [
-      'img/apartment-03.jpg',
-      'img/studio-01.jpg',
-      'img/room.jpg',
-      'img/apartment-02.jpg',
-      'img/apartment-01.jpg',
-    ],
-    rating: 3.9,
-    bedrooms: 3,
-    maxAdults: 4,
-    goods: [
-      Good.TOWELS,
-      Good.FRIDGE,
-      Good.CABEL_TV,
-      Good.BABY_SEAT,
-      Good.HEATING,
-    ],
-    host: {
-      id: 23,
-      name: HostName.JAMES,
-      isPro: false,
-      avatarUrl: '',
-    },
-    city: City[CityName.HAMBURG],
-    isPremium: false,
-    isFavourite: false,
-    location: {
-      latitude: 53.555,
-      longitude: 9.984,
-      zoom: ZOOM_OFFER,
-    },
-  },
-  {
-    id: 'offer008',
-    title: 'Cosy double room',
-    description: 'Kick your journey off with a coffee on the 8th floor rooftop to take on its typical pretty Parisian view and finish it with a fresh pint in the bar and its patio. Sleep tight in our brand new comfy beds. Repeat.',
-    type: OfferType.APARTMENT,
-    price: 199,
-    previewImage: 'img/apartment-02.jpg',
-    images: [
-      'img/apartment-02.jpg',
-      'img/apartment-01.jpg',
-      'img/apartment-03.jpg',
-      'img/studio-01.jpg',
-    ],
-    rating: 4.8,
-    bedrooms: 2,
-    maxAdults: 3,
-    goods: [
-      Good.WI_FI,
-      Good.TOWELS,
-      Good.FRIDGE,
-      Good.CABEL_TV,
-      Good.HEATING,
-    ],
-    host: {
-      id: 23,
-      name: HostName.JAMES,
-      isPro: false,
-      avatarUrl: '',
-    },
-    city: City[CityName.PARIS],
-    isPremium: false,
-    isFavourite: true,
-    location: {
-      latitude: 48.864716,
-      longitude: 2.349014,
-      zoom: ZOOM_OFFER,
-    },
-  },
-  {
-    id: 'offer009',
-    title: 'Magnificent views of the river and surroundings',
-    description: 'All sights of the city center are within walking distance. Sleep tight in our brand new comfy beds. Repeat.',
-    type: OfferType.PRIVATE_ROOM,
-    price: 99,
-    previewImage: 'img/apartment-01.jpg',
-    images: [
-      'img/apartment-02.jpg',
-      'img/apartment-01.jpg',
-      'img/apartment-03.jpg',
-      'img/studio-01.jpg',
-    ],
-    rating: 3.6,
-    bedrooms: 2,
-    maxAdults: 3,
-    goods: [
-      Good.WI_FI,
-      Good.TOWELS,
-      Good.HEATING,
-    ],
-    host: {
-      id: 23,
-      name: HostName.JAMES,
-      isPro: false,
-      avatarUrl: '',
-    },
-    city: City[CityName.PARIS],
-    isPremium: false,
-    isFavourite: false,
-    location: {
-      latitude: 48.866,
-      longitude: 2.354,
-      zoom: ZOOM_OFFER,
-    },
-  },
-  {
-    id: 'offer010',
-    title: 'Excellent accommodation conditions',
-    description: 'If you have long wanted a leisurely vacation and a wonderful experience, then you should definitely pay attention to this option.',
-    type: OfferType.APARTMENT,
-    price: 90,
-    previewImage: 'img/apartment-03.jpg',
-    images: [
-      'img/apartment-02.jpg',
-      'img/apartment-01.jpg',
-      'img/apartment-03.jpg',
-      'img/studio-01.jpg',
-    ],
-    rating: 4.5,
-    bedrooms: 1,
-    maxAdults: 2,
-    goods: [
-      Good.WI_FI,
-      Good.TOWELS,
-      Good.FRIDGE,
-      Good.CABEL_TV,
-      Good.HEATING,
-    ],
-    host: {
-      id: 23,
-      name: HostName.JAMES,
-      isPro: false,
-      avatarUrl: '',
-    },
-    city: City[CityName.PARIS],
-    isPremium: true,
-    isFavourite: true,
-    location: {
-      latitude: 48.864,
-      longitude: 2.33,
-      zoom: ZOOM_OFFER,
-    },
-  },
-  {
-    id: 'offer011',
-    title: 'Altbauwohnung im Herzen von Ehrenfeld!',
-    description: 'Herzlich Willkommen in meinem 1,5 Zimmer Appartement im beliebten Ehrenfeld. Das Appartement liegt im Herzen Ehrenfeld. Bars, Bahn, Einkaufsmöglichkeiten liegen fußläufig entfernt. Wer Köln kennenlernen will, ist hier genau richtig!',
-    type: OfferType.PRIVATE_ROOM,
-    price: 79,
-    previewImage: 'img/apartment-01.jpg',
-    images: [
-      'img/apartment-01.jpg',
-      'img/apartment-02.jpg',
-      'img/apartment-03.jpg',
-    ],
-    rating: 2.7,
-    bedrooms: 1,
-    maxAdults: 2,
-    goods: [
-      Good.WI_FI,
-      Good.TOWELS,
-      Good.HEATING,
-    ],
-    host: {
-      id: 23,
-      name: HostName.JAMES,
-      isPro: false,
-      avatarUrl: '',
-    },
-    city: City[CityName.COLOGNE],
-    isPremium: false,
-    isFavourite: false,
-    location: {
-      latitude: 50.935173,
-      longitude: 6.953101,
-      zoom: ZOOM_OFFER,
-    },
-  },
-  {
-    id: 'offer012',
-    title: 'Ausgezeichnete Unterkunftsbedingungen',
-    description: 'Herzlich Willkommen in meinem 1,5 Zimmer Appartement im beliebten Ehrenfeld. Das Appartement liegt im Herzen Ehrenfeld. Bars, Bahn, Einkaufsmöglichkeiten liegen fußläufig entfernt. Wer Köln kennenlernen will, ist hier genau richtig!',
-    type: OfferType.APARTMENT,
-    price: 100,
-    previewImage: 'img/apartment-02.jpg',
-    images: [
-      'img/apartment-01.jpg',
-      'img/apartment-02.jpg',
-      'img/apartment-03.jpg',
-    ],
-    rating: 4.4,
-    bedrooms: 1,
-    maxAdults: 2,
-    goods: [
-      Good.WI_FI,
-      Good.TOWELS,
-      Good.HEATING,
-    ],
-    host: {
-      id: 23,
-      name: HostName.JAMES,
-      isPro: false,
-      avatarUrl: '',
-    },
-    city: City[CityName.COLOGNE],
-    isPremium: false,
-    isFavourite: false,
-    location: {
-      latitude: 50.938,
-      longitude: 6.943,
-      zoom: ZOOM_OFFER,
-    },
-  },
-  {
-    id: 'offer013',
-    title: 'Cosy red-brick studio',
-    description: 'This warm and cosy red-bricked studio is located just a stone\'s throw from the well-known Grand-Place and Bourse, in the very heart of Brussels.',
-    type: OfferType.APARTMENT,
-    price: 100,
-    previewImage: 'img/apartment-02.jpg',
-    images: [
-      'img/apartment-01.jpg',
-      'img/apartment-02.jpg',
-      'img/apartment-03.jpg',
-    ],
-    rating: 4.7,
-    bedrooms: 1,
-    maxAdults: 2,
-    goods: [
-      Good.WI_FI,
-      Good.HEATING,
-      Good.FRIDGE,
-      Good.CABEL_TV,
-    ],
-    host: {
-      id: 23,
-      name: HostName.JAMES,
-      isPro: false,
-      avatarUrl: '',
-    },
-    city: City[CityName.BRUSSELS],
-    isPremium: false,
-    isFavourite: false,
-    location: {
-      latitude: 50.8505,
-      longitude: 4.3488,
-      zoom: ZOOM_OFFER,
-    },
-  },
-  {
-    id: 'offer014',
-    title: 'Room with a wonderful view',
-    description: 'This warm and cosy red-bricked studio is located just a stone\'s throw from the well-known Grand-Place and Bourse, in the very heart of Brussels.',
-    type: OfferType.PRIVATE_ROOM,
-    price: 130,
-    previewImage: 'img/room.jpg',
-    images: [
-      'img/studio-01.jpg',
-      'img/room.jpg',
-    ],
-    rating: 3.2,
-    bedrooms: 1,
-    maxAdults: 3,
-    goods: [
-      Good.WI_FI,
-      Good.HEATING,
-      Good.CABEL_TV,
-    ],
-    host: {
-      id: 23,
-      name: HostName.JAMES,
-      isPro: false,
-      avatarUrl: '',
-    },
-    city: City[CityName.BRUSSELS],
-    isPremium: false,
-    isFavourite: false,
-    location: {
-      latitude: 50.854,
-      longitude: 4.345,
-      zoom: ZOOM_OFFER,
-    },
-  },
-  {
-    id: 'offer015',
-    title: 'Exclusive apartment directly in the city center',
-    description: 'Modern 40sm² apartment with living room/sleeping room, separate kitchen, and bathroom only 5 min by feet from the city centre, airport and fair/exhibition only 10 minutes by taxi.',
-    type: OfferType.APARTMENT,
-    price: 50,
-    previewImage: 'img/apartment-03.jpg',
-    images: [
-      'img/apartment-01.jpg',
-      'img/apartment-02.jpg',
-      'img/apartment-03.jpg',
-    ],
-    rating: 1.7,
-    bedrooms: 1,
-    maxAdults: 2,
-    goods: [
-      Good.WI_FI,
-      Good.FRIDGE,
-      Good.CABEL_TV,
-    ],
-    host: {
-      id: 23,
-      name: HostName.JAMES,
-      isPro: false,
-      avatarUrl: '',
-    },
-    city: City[CityName.DUSSELDORF],
-    isPremium: true,
-    isFavourite: true,
-    location: {
-      latitude: 51.233334,
-      longitude: 6.783333,
-      zoom: ZOOM_OFFER,
-    },
-  },
-  {
-    id: 'offer016',
-    title: 'Luxurious apartment at great location',
-    description: 'All commodities, jacuzzi and beautiful scenery. Ideal for families or friends. Only 5 min by feet from the city centre, airport and fair/exhibition only 10 minutes by taxi.',
-    type: OfferType.APARTMENT,
-    price: 90,
-    previewImage: 'img/room.jpg',
-    images: [
-      'img/room.jpg',
-      'img/apartment-01.jpg',
-      'img/apartment-02.jpg',
-      'img/apartment-03.jpg',
-    ],
-    rating: 1.7,
-    bedrooms: 1,
-    maxAdults: 2,
-    goods: [
-      Good.WI_FI,
-      Good.FRIDGE,
-      Good.CABEL_TV,
-    ],
-    host: {
-      id: 23,
-      name: HostName.JAMES,
-      isPro: false,
-      avatarUrl: '',
-    },
-    city: City[CityName.DUSSELDORF],
-    isPremium: false,
-    isFavourite: false,
-    location: {
-      latitude: 51.236,
-      longitude: 6.788,
-      zoom: ZOOM_OFFER,
-    },
-  },
-  {
-    id: 'offer018',
-    title: 'Best place to stay in the city',
-    description: 'Only 5 min by feet from the city centre, airport and fair/exhibition only 10 minutes by taxi.',
-    type: OfferType.APARTMENT,
-    price: 190,
-    previewImage: 'img/apartment-01.jpg',
-    images: [
-      'img/room.jpg',
-      'img/apartment-01.jpg',
-      'img/apartment-02.jpg',
-      'img/apartment-03.jpg',
-    ],
-    rating: 4.7,
-    bedrooms: 1,
-    maxAdults: 2,
-    goods: [
-      Good.WI_FI,
-      Good.FRIDGE,
-      Good.BABY_SEAT,
-      Good.CABEL_TV,
-      Good.COFFEE_MASHINE,
-      Good.DISHWASHER,
-    ],
-    host: {
-      id: 23,
-      name: HostName.JAMES,
-      isPro: false,
-      avatarUrl: '',
-    },
-    city: City[CityName.DUSSELDORF],
-    isPremium: false,
-    isFavourite: false,
-    location: {
-      latitude: 51.231,
-      longitude: 6.793,
-      zoom: ZOOM_OFFER,
-    },
-  },
-];
-
-export default offers;
