+++ project/src/prop-types/city.prop.js
@@ -1,5 +1,5 @@
 import PropTypes from 'prop-types';
-import {CityName} from '../mocks/constants';
+import {CityName} from '../constants.js';
 
 export default PropTypes.oneOf(
   [
