+++ project/src/settings.js
@@ -1,4 +1,4 @@
-import {CityName} from './mocks/constants.js';
+import {CityName} from './constants.js';
 
 const RATING_STARS = 5;
 const DEFAULT_CITY = CityName.PARIS;
