+++ project/src/store/reducer.js
@@ -4,7 +4,6 @@ import {ActionType} from './action.js';
 const initialState = {
   city: DEFAULT_CITY,
   offers: [],
-  realOffers: [],
   reviews: [],
 };
 
@@ -15,17 +14,11 @@ const reducer = (state = initialState, action) => {
         ...state,
         city: action.payload,
       };
-    case ActionType.GET_OFFERS:
+    case ActionType.LOAD_OFFERS:
       return {
         ...state,
         offers: action.payload,
       };
-    case ActionType.LOAD_OFFERS:
-      return {
-        ...state,
-        realOffers: action.payload,
-      };
-
     default:
       return state;
   }
